requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.7.1
astroid==2.15.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
cachelib==0.10.2
cachetools==5.2.0
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.6
coverage==7.2.3
cssselect==1.2.0
cycler==0.11.0
dataclasses-json==0.5.9
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
distro==1.8.0
dnspython==2.3.0
docker==6.0.1
docopt==0.6.2
duckduckgo-search==2.8.5
et-xmlfile==1.1.0
exceptiongroup==1.0.4
executing==1.2.0
filelock==3.11.0
flake8==6.0.0
Flask==2.3.2
Flask-PyMongo==2.3.0
Flask-Session==0.5.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
gTTS==2.3.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.1
huggingface-hub==0.13.4
identify==2.5.22
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
ipykernel==6.23.2
ipython==8.14.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jira==3.5.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.1
kiwisolver==1.4.4
langchain==0.0.218
langchainplus-sdk==0.0.20
lazy-object-proxy==1.9.0
loguru==0.7.0
lxml==4.9.2
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
nodeenv==1.7.0
numexpr==2.8.4
numpy==1.24.0
oauthlib==3.2.2
openai==1.6.1
openapi-schema-pydantic==1.2.4
openpyxl==3.0.9
orjson==3.8.10
outcome==1.2.0
packaging==23.1
pandas==1.5.2
pandas-stubs==1.2.0.35
parso==0.8.3
pathspec==0.11.1
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.3.0
pinecone-client==2.2.1
pipreqs==0.4.13
platformdirs==3.2.0
playsound==1.2.2
pluggy==1.2.0
pre-commit==3.2.2
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
pure-eval==0.2.2
py2exe==0.13.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.11
pyflakes==3.0.1
Pygments==2.15.1
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pylint==2.17.2
pymongo==4.4.1
pyodbc==4.0.39
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==25.1.0
readability-lxml==0.8.1
redis==4.5.4
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.1
selenium==4.7.2
sentence-transformers==2.2.2
sentencepiece==0.1.98
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sourcery==1.2.0
SQLAlchemy==2.0.18
stack-data==0.6.2
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
trio==0.22.0
trio-websocket==0.9.2
typing-inspect==0.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
virtualenv==20.21.0
wcwidth==0.2.6
webdriver-manager==3.8.5
websocket-client==1.5.1
Werkzeug==2.3.6
win32-setctime==1.1.0
wrapt==1.15.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
